ARM: dts: aspeed: romulus: config gpio mux for adc battery
When GPIOR3 is pulled high for ADC to read battery voltage, it drains
the battery as well.
So config a gpio mux with io-channel-mux for this channel, and set the
gpio mux's idle-state to 0 to save battery.

With the config, when reading the battery voltage, kernel will
automatically set 1 to GPIOR3, and after reading is done, kernel will
set 0.

ADC channel expects some time to sample the voltage, so added 1ms delay
before reading raw value in aspeed adc driver, which is suggested by
Aspeed.

Signed-off-by: Lei YU <mine260309@gmail.com>